refactor: 중복 코드 제거를 위한 유틸 클래스 네이밍 및 디렉토리 수정
- ReservationConverter -> JdbcRemoveDuplicateUtils
- ReservationConverter.get() -> JdbcRemoveDuplicateUtils.getReservationFromResultSet()
- ReservationConverter.set() -> JdbcRemoveDuplicateUtils.setReservationToStatement()
- util 디렉토리로 이동
- 세부적인 함수명으로 변경
